renamed Comment.tsx
